requirements.txt
mcp>=1.0.0
ddddocr>=1.4.0
httpx>=0.27.0
pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
x402[evm,fastapi]>=2.8.0
eth-account>=0.13.0
pydantic>=2.0.0
